Moved socketio code from main to socketioMethods
